Move goto_symex_is_constantt into its own file
No behavioural change.
